controllers & repos now complete, exercise abstract class to be potentially deleted
